+++ src/main/java/com/jsnunez/peliculas/domain/repository/AlmacenRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Almacen;
 
 public interface AlmacenRepository extends
-CrudRepository<Almacen,Byte> {
+CrudRepository<Almacen,Long> {
 }
