+++ src/main/java/com/jsnunez/peliculas/domain/repository/CategoriaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Categoria;
 
 public interface CategoriaRepository  extends
-CrudRepository<Categoria,Short> {
+CrudRepository<Categoria,Long> {
 }
