+++ src/main/java/com/jsnunez/peliculas/domain/repository/CiudadRepository.java
@@ -3,5 +3,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jsnunez.peliculas.domain.entities.Ciudad;
 public interface CiudadRepository extends
-CrudRepository<Ciudad,Short> {
+CrudRepository<Ciudad,Long> {
 }
