+++ src/main/java/com/jsnunez/peliculas/domain/repository/DireccionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Direccion;
 
 public interface DireccionRepository  extends
-CrudRepository<Direccion,Short> {
+CrudRepository<Direccion,Long> {
 }
