+++ src/main/java/com/jsnunez/peliculas/domain/repository/EmpleadoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Empleado;
 
 public interface EmpleadoRepository  extends
-CrudRepository<Empleado,Byte> {
+CrudRepository<Empleado,Long> {
 }
