+++ src/main/java/com/jsnunez/peliculas/domain/repository/IdiomaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Idioma;
 
 public interface IdiomaRepository extends
-CrudRepository<Idioma,Byte> {
+CrudRepository<Idioma,Long> {
 }
