+++ src/main/java/com/jsnunez/peliculas/domain/repository/PaisRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.jsnunez.peliculas.domain.entities.Pais;
 
 public interface PaisRepository  extends
-CrudRepository<Pais,Short> {
+CrudRepository<Pais,Long> {
 }
