+++ src/main/java/com/jsnunez/peliculas/domain/service/DireccionService.java
@@ -8,7 +8,7 @@ public interface DireccionService   {
 
     
     List<Direccion> findAll();
-Optional<Direccion> findById(short id);
+Optional<Direccion> findById(Long id);
 Direccion save(Direccion Direccion);
 Direccion update(Long id);
 Optional<Direccion> delete(Long id);
