+++ src/main/java/com/jsnunez/peliculas/domain/service/PaisService.java
@@ -7,7 +7,7 @@ import com.jsnunez.peliculas.domain.entities.Pais;
 public interface PaisService {
     
     List<Pais> findAll();
-Optional<Pais> findById(short id);
+Optional<Pais> findById(Long id);
 Pais save(Pais Pais);
 Pais update(Long id);
 Optional<Pais> delete(Long id);
